pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.0.3",
]
